backend/requirements/dev.txt
black==19.10b0
django-debug-toolbar==2.2
watchdog==0.10.3
pyyaml==5.3.1
argh==0.26.2
django-extensions==3.0.5
ipython==7.17.0
jupyter==1.0.0
Werkzeug==1.0.1
pylint-django==2.3.0